requirements.txt
timm>=0.9.16
diffusers==0.26.3
transformers==4.38.2
open_clip_torch==2.24.0
accelerate==0.27.2
kornia
omegaconf
lpips
matplotlib
opencv-python
wandb
segment-anything-hq
grad-cam

